Write book page + React Quill integration
I have built out the design for the WriteBook page to be inline with the rest of the apps design.

I have added the React Quill New library to the project as well for the actual story writing box, as it gives some great features for users writing their stories.

Small refactors here and there and minor style adjustments.