Functionality for creating a story
I have built out the main functionality for the writing a story page. There are still some features missing but they will be added later.

I have use firebase for my db solution so most of this is just interacting with the firebase API to add documents to firebase to store the stories.

I have also ensured that each story contains the userId of the creator and vice versa i have saved the storyId to the creator's stories array on the db. This will enable me later to have a page that will show all the users stories and allow them to edit/delete them.

I am also now capturing a full name from the user to display on the story later when showing them on other pages.

Some minor refactors/optimizations.