Library page
Initially we are going to have the user select a genre, this should firstly reduce our db load as we will need to pull less data but also will help me use react router and the url to get the correct data to show on the next page which will be a list that will show all stories of the selected genre.

I have designed the page exactly as i want it which has meant hardcoding some values where needed.  to get the exact look i wanted.

I have use a .map to generate my cards on the fly and have added some code to color and provide images to the relevant cards. This is quite scalable as adding another genre simply means adding one to the genres array and creating new aligns and spans. I wanted the spans to be exact but you could randomly generate these to achieve a similar effect which would be a little more scalable but wasnt suitable for this project.