React router params
Added react router links to the genre cards, this will enable me to store the genre in the params to be retrieved in the story list component and used for the pulling of the stories of that genre.

Added a button to fill some empty space although the user can click the card or the button it makes no difference.

Minor styling changes and additions.