Structuring future pages
I am still in the designing phase of these components and i am searching for a way to import a bunch of short stories such as an API to atleast have some data on there, i may even just go with the AI generated sample data i have put into the stories.js file. They will work for the building but is definately a sub-optimal solution.

Other than that i have just laid out some basic structure for the comps i know i will need. There are still a decent few more needed but we will need the data for most of them.