Story List sample data and layout
Added fakerJS to allow me to spoof some data so that there can actually be some "stories" there to look at. Although this does mean its completely nonsensical i think it will serve my purposes. Atleast for now.

Added the layout for the storyList page as well as the header which is complete. Styling is still needed here.

Added storycard component which will serve as a link to take the user from the library to the book page where they can then read it.