Story cover images
I have added images for the book covers. I was originally going to leave this out as my storage solution no longer allows image storage for free so for adding users own covers it wont work but i can assign a random image to give the illusion of one.

There seems to be an issue with fakers random images where up to half of the images it tries to pull are non-existent resulting in an empty image box. Ive implemented to me what feels like a hacky way around it by implementing a function to check the validity of the image its trying to pull, if it works it sends the url to the frontend, if not it re runs the fetch request. Im not sure how performant this is but faker does not allow me to do this before fetching the image sadly it seems a little antiquated. But this does work for now.