Placeholder images
I have noticed that the images from faker take a while to paint to the screen. This is an unfortunate result of using faker. I thought about adding a loading spinner, but some of them take a good while to come through so i have opted for place holder images instead.

In the real world these images would be retrieved from the back end database and served to the front end immediately but here sadly we have to make a call to the API and in this case multiple as half the time faker will return a dead url resulting in a blank image.