Data Seeding
I have been really thinking about how i want to handle the faker sample data implementation because i want the site to be fully functional and integrated with the firebase database. So that if a user does create a short story such as a respective employer they will be able to do so and have it shown on the storyList for anyone to see. To pull data from the db and create faker sample data on each reload seems less than optimal. So i have written 2 solutions and which i will use i still dont know yet.

The first is to simply generate the faker data once and store it in a stories array  or context and simply filter this array based on the genre the user is looking for.

The second is to seed the firebase database with the faker data once and then solely interface with firebase as i would with a real world application. This feels like the better option but isnt something ive done before so im wary of doing this and potentially filling my db with duplicate or excessive data. So for now ive written both while i deliberate what to do.