Database Seeding
So i figured that seeding the database would be the best option and just work solely from that. I also found out about the fact that picsum does seeded images allowing me to circumvent my validation altogether as it will never return a broken image link. I wish i had known about it sooner but thank god i know it now. With this puzzle piece and some trial and error and much testing with single genres etc i was able to get this working and the database is now seeded with 15 stories for each genre and any user generated ones will simply be added to this by interfacing with firebase. An elegant solution in my mind.

I have also done some minor refactoring to get everything in line with this new approach and drop the readTime in favour of an image etc.