Book Page
I have added a page to open up and read the books.

The seed data has the issue of not being styled unfortunately, as its generated by faker. So what i have done is added another script to go through and add an isSeedData boolean to each of the seeded stories so that if we find its a styled one e.g. from a user then we will use purified html to render it as the user styled it. If we find that the boolean is true and its a none styled one we render a simpler component to just repeat the text to fill the page.