My Stories List
I have added a My Stories page. I do now feel that this will become a component on the Account page. Its a simple refactor so i wont worry about it for now.

This list pulls all stories from the database where the logged in user's id === to the creatorID of a story. This way we know we are only pulling the stories this user is allowed to edit.

It then renders these stories in a list where the user can edit, delete and change who can see the story(either just the writer or everyone).

I still need to make these buttons functional. Which thanks to firebase shouldnt be too hard.

Minor refactors and style changes.