Deleting Stories
Added the ability for a user to delete their own stories.

Added a modal window to confirm the user wants to delete it and makes sure they are aware it is irreversible before then running the handle delete function to remove it from the db and refresh to component.

Added navigation once a book has been written the user will be navigated to the book page.

Added disabled prop for writebook inputs when a book is being posted.

Fixed error with styling whitespace in book header component.

Fixed bug with Science Fiction genre not correctly manipulating the genre string.

Minor styling changes.