Homepage
Added the hompage functionality to make it so that it pulls a random story from the db and shows it as a suggested read. I might limit this to just a couple of genre's to make it less expensive as right now it pulls the entire array of stories just to use 1 of them which doesnt feel optimal.