Commenting
Added the ability for users to comment on stories. I used my updateStories script to add an empty comments array to each story for the sake of consistency.

The comments component will pull all comments for each story. Which i am saving the comments array to the story itself so all we need to get them is the storyId. Then it generates a commentCard per comment.

I still need to add the ability to delete a comment however.