Liking stories
I have added a likes system, allowing users to like stories they enjoyed and i will later add the ability for users to browse their likes which will allow functionality for saving stories. I could extract this into its own favorite component or read later which i might do I'm unsure at the moment. I also refactored the way the story is pulled in this component to make sure a re-fetch is triggered when the db is altered resulting in instant re-renders and visual feedback to the user that they have liked it.