Latest story page
Added a latest story page to the homepage. Now the issue here is that if i code it to actually show the story with the most recent createdAt value it will always be my story and not one of the example seed data stories which isnt what i want so instead it pulls a random story.