Favorite
Added functionality for a user saving a story as a favorite. Theres no where yet to read them back as im not sure where i want to put it yet.

The user clicks favorite on a story and the id of that story is then added to the favorite array of that user that we are saving in the users collection. Which is one of the reasons when we create a user we also add that user to a firestore collection called users.

Minor styled changes/refactors