Account page and favorites
I have decided i am going to eventually refactor the mystories page into a component for render here in the accounts page. For now its set up with router and its own navigation to render the desired component.

Added favorites component to be rendered in this outlet. A user can now favorite a story which can also serve as a read later list as they can view favorites at any time and unfavorite it from this component as well as open it to be read.

Minor bug fixes/optimizations and styling changes.