Edit account details
Added the ability to edit account details, which in this case consists of only the fullName and displayName. I have diabled the email and password inputs as firebase uses real email verification to change email and passwords and as this is just a portfolio piece i dont expect users to use a real email so i have disabled this.