Protected Routes + Optimizations
I have added protected routes to make sure we have a user logged in before allowing them to do things such as writing stories liking, favorites, etc.

I have also added some fixes and optimizations such as making sure that on the book we display the correct display name for the author, i noticed if the user changes their display name it would not follow suit on the book page or the story card, which i do not want. If the user changes their display name it should change on their books as well. I have simply added a useEffect to each of these components to get the display name from the db using the story's creatorID. This keeps everything in sync nicely so if a user changes their display name it will automatically change it on these components now.