Homepage Refactor
I have refactored the homepage so that instead of pulling all of the stories from the back end and locally choosing 3 random ones i have used my updateStories script and given them all a random number which we will also do for any new stories and then use this number to only pull the 5 we need for the homepage. This is a lot more performant and scalable.

Minor styling changes and optimizations.