Sorting and Searching
Added sorting functionality to the story list allowing user to sort it based on when it was created or which has the most likes, which in this case i feel is all that is needed. I have also added search functionality allowing the user to search for any book using its title or author.

I have also fixed an issue where there were multiple sources of truth for the displayName which i have now corrected to make it so we always take the displayName from the firestore. I have also ensured however that the profile is updated with the new displayName even though we will never read it from there its good to keep it accurate.