Form validation
I have added form validation. This ensures that users arent using prohibited characters for the use of injection such as "<". I have also ensured the required input fields are filled and that they meet minimum and maximum length requirements.

Refactored and removed unnecessary code such as forcing re-renders on delete as we are now using onSnapshot the data is re-fetched following any changes to the backend resulting in live updates to the front-end.

Found and fixed an issue where the storyId was not being removed from the user's stories array when a story was deleted.

I have also added react-toastify which i will be using for user notifications for submissions and error handling.