Error handling
Added page not found styling and functionality for when a user tries to navigate to a broken or missing link.

Added error handling in the components that need it. I believe i got all of them done. This was somewhat more complicated than i expected as this is my first project in a little while im a bit out of practice but i believe i have accounted for both component errors and page errors and i will either render the error component or give the user a toast notification to inform them of the issue.