Styling loading spinner / navbar
I have refactored my navbar to make use of the NavLinks we get from react router so that it is both more semantic and we can give them a  class based on if its the active link.

I have added a loading spinner where needed and given it a custom $height prop to allow us to determine its size and positioning as it will vary depending on where its used.

Some refactors and visual changes/updates