Saving as drafts
Added ability to save stories as drafts. This will not publish the story. These drafts can then be edited by the user and posted as an actual story whenever they like, they can also update and save the draft as much as they like before posting. When a user does post a draft as a story the draft is removed from the users drafts array.

Fixed bug with custom input component not being properly controlled.

Refactored EditStory component.

Minor styling fixes and changes.