Mobile Navbar
I have added a responsive mobile and tablet burger menu that replaces the navbar at the breakpoint i found that the navbar styling starts to break. I have used framer motion to animate the navbar and i will also be using it to animate other parts of the site - this gave me some grief as i was encountering really strange issues with the conditional renders and the staggerChildren propagation the amount of links where the animation was applied would increase on each refresh but not to all of them, this turned out to be because the links were not available immediately on mount when the animation was trying to work meaning it would only apply to the last 3 as they would be available in time whereas the first 3 would not.