MyStories / Account Responsive
I have now made the my stories page and the accounts page fully responsive down to max-width: 300px.

This gave me considerable grief to get working and looking clean but i believe its gotten there in the end.

I have also fixed the fact that the "mystories" "favorites" and "drafts" lists where scrollable as well as the page. I have removed that and instead opted for the page to grow to fit the content.