Account navbar bug fix
Fixed a bug where the dropdown list on the mobile layout version of the account navbar was not correctly synced with the current route. So there was some strange behavior with what link was shown in the dropdown and it would always default to favorites even if we are on a different page. This is now fixed.