Featured banner functionality
I have the explore featured stories button work. I have made it scroll to the staff picks section. Due to the navbar being out of flow i need to account for its height and make sure not to over scroll by the navHeight. I then also have made it so this logic doesn't apply on the mobile layout when the navbar is just a burger menu.