Sign Up redesign and responsiveness.
I have redesigned the signup page as when i went into it to make it responsive it felt a bit lack luster, so i have made it more inline with the design of the sign in page and i think it looks a lot better now.

I have also made it responsive down to 300px viewport width as well.