Scalability Improvements
It occurred to me that right now this app only uses around 120ish stories, which is a small size really. But as its purpose is to eventually take a potentially infinite number of user generated content i felt it prudent to plan for this eventuality as my previous implementation was pulling all the stories for a given genre, which is fine when each genre only has 15 stories, but the more and more we get the worse it is. Tackling this gave me some grief as it became apparent that my sorting would no longer work if i were to just pull 15 results at a time. Thankfully after some research i found out that firebase has great options for handling the sorting logic on the server-side, so now we only pull the results we need and paginate the rest. Doing the same for the likes sorting was also an issue and to fix it i have had to add a likesCount to every story to enable performant server-side sorting based on the likes and i am also now making sure this number is incremented when a user likes or removes their like.

The server side-sorting was also encountering issues as not every story had a hidden property as that feature was added later which was fine because in my code where i was using the property hidden = undefined is falsey and so it worked fine, on the server side though this wasnt the case and it was pulling none of the stories until i added a hidden property to every story.

I also encountered an issue with

Fixed an issue where the make public and make private buttons had inverted  tooltips saying make private when clicking it would actually make it public.