Mobile layout buttons fix
The icons i was using for the buttons are not very accessible to mobile users as they require a hover to see the tooltip. To fix this i have replaced them with actual buttons on smaller viewports on the mystories, favorites and drafts pages.

Minor styling changes and fixe.

Minor comment cleanup.